Upload my version of __init__
It looks like the main differences are where we do 'neighbors = mol.get_neighbors...' and then loop over the list.  There are a few places where the choice of tabs or spaces was inconsistent.  Also, for some reason we chose different formats for printing to the Bq file (15.8f or 15.6f).